+++ src/build.task.ts
@@ -6,9 +6,9 @@ import type {Task, Args} from './task/task.js';
 import type {Map_Input_Options, Map_Output_Options, Map_Watch_Options} from './build/rollup.js';
 import {load_config} from './config/config.js';
 import type {Gro_Config} from './config/config.js';
-import type {Task_Events as Server_Task_Events} from './server.task.js';
 import type {Adapter_Context, Adapter} from './adapt/adapter.js';
 import {build_source} from './build/build_source.js';
+import type {Plugin, Plugin_Context} from './plugin/plugin.js';
 
 export interface Task_Args extends Args {
 	map_input_options?: Map_Input_Options;
@@ -16,7 +16,7 @@ export interface Task_Args extends Args {
 	map_watch_options?: Map_Watch_Options;
 }
 
-export interface Task_Events extends Server_Task_Events {
+export interface Task_Events {
 	'build.create_config': (config: Gro_Config) => void;
 	'build.build_src': void;
 }
@@ -37,12 +37,44 @@ export const task: Task<Task_Args, Task_Events> = {
 		timing_to_load_config();
 		events.emit('build.create_config', config);
 
+		// Create the production plugins
+		// TODO this has a lot of copypaste with `gro dev` plugin usage,
+		// probably extract a common interface
+		const timing_to_create_plugins = timings.start('create plugins');
+		const plugin_context: Plugin_Context<Task_Args, Task_Events> = {
+			...ctx,
+			config,
+			filer: null,
+		};
+		const plugins: Plugin<any, any>[] = to_array(await config.plugin(plugin_context)).filter(
+			Boolean,
+		) as Plugin<any, any>[];
+		timing_to_create_plugins();
+
 		// Build everything with esbuild and Gro's `Filer` first.
 		// These production artifacts are then available to all adapters.
 		const timing_to_build_src = timings.start('build_src');
 		await build_source(fs, config, dev, log);
 		timing_to_build_src();
 		events.emit('build.build_src');
+
+		const timing_to_call_plugin_setup = timings.start('setup plugins');
+		for (const plugin of plugins) {
+			if (!plugin.setup) continue;
+			const timing = timings.start(`setup:${plugin.name}`);
+			await plugin.setup(plugin_context);
+			timing();
+		}
+		timing_to_call_plugin_setup();
+
+		const timing_to_call_plugin_teardown = timings.start('teardown plugins');
+		for (const plugin of plugins) {
+			if (!plugin.teardown) continue;
+			const timing = timings.start(`teardown:${plugin.name}`);
+			await plugin.teardown(plugin_context);
+			timing();
+		}
+		timing_to_call_plugin_teardown();
 
 		// Adapt the build to final ouputs.
 		const timing_to_create_adapters = timings.start('create adapters');
@@ -56,15 +88,6 @@ export const task: Task<Task_Args, Task_Events> = {
 		timing_to_create_adapters();
 
 		if (adapters.length) {
-			const timing_to_call_begin = timings.start('begin');
-			for (const adapter of adapters) {
-				if (!adapter.begin) continue;
-				const timing = timings.start(`begin:${adapter.name}`);
-				await adapter.begin(adapt_context);
-				timing();
-			}
-			timing_to_call_begin();
-
 			const timing_to_call_adapt = timings.start('adapt');
 			for (const adapter of adapters) {
 				if (!adapter.adapt) continue;
@@ -73,15 +96,6 @@ export const task: Task<Task_Args, Task_Events> = {
 				timing();
 			}
 			timing_to_call_adapt();
-
-			const timing_to_call_end = timings.start('end');
-			for (const adapter of adapters) {
-				if (!adapter.end) continue;
-				const timing = timings.start(`end:${adapter.name}`);
-				await adapter.end(adapt_context);
-				timing();
-			}
-			timing_to_call_end();
 		} else {
 			log.info('no adapters to `adapt`');
 		}