+++ src/config/gro.config.default.ts
@@ -5,12 +5,14 @@ import {
 	has_node_library,
 	NODE_LIBRARY_BUILD_CONFIG,
 	has_sveltekit_frontend,
+	has_api_server,
+	API_SERVER_BUILD_CONFIG,
 } from '../build/default_build_config.js';
 
 /*
 
-This is the default config that's used
-if the current project does not define one at `src/gro.config.ts`.
+This is the default config that's passed to `src/gro.config.ts`
+if it exists in the current project, and if not, this is the final config.
 It looks at the project and tries to do the right thing:
 
 - if `src/routes` and `src/app.html`,
@@ -27,30 +29,41 @@ It looks at the project and tries to do the right thing:
 export const config: Gro_Config_Creator = async ({fs}) => {
 	const [
 		enable_node_library,
-		// enable_api_server,
+		enable_api_server,
 		enable_sveltekit_frontend,
 		// enable_gro_frontend,
 	] = await Promise.all([
 		has_node_library(fs),
-		// has_api_server(fs),
+		has_api_server(fs),
 		has_sveltekit_frontend(fs),
 		// has_deprecated_gro_frontend(fs),
 	]);
 	const partial: Gro_Config_Partial = {
 		builds: [
 			enable_node_library ? NODE_LIBRARY_BUILD_CONFIG : null,
-			// enable_api_server ? API_SERVER_BUILD_CONFIG : null,
+			enable_api_server ? API_SERVER_BUILD_CONFIG : null,
 			// enable_gro_frontend ? toDefaultBrowserBuild() : null, // TODO configure asset paths
 			// note there's no build for SvelteKit frontends - should there be?
 		],
 		log_level: ENV_LOG_LEVEL ?? Log_Level.Trace,
 		types: enable_node_library,
+		plugin: async () => [
+			enable_api_server
+				? (await import('../plugin/gro-plugin-api-server.js')).create_plugin()
+				: null,
+			enable_sveltekit_frontend
+				? (await import('../plugin/gro-plugin-sveltekit-frontend.js')).create_plugin()
+				: null,
+		],
 		adapt: async () => [
 			enable_node_library
 				? (await import('../adapt/gro-adapter-node-library.js')).create_adapter()
 				: null,
-			// TODO
-			// enable_api_server ? (await import('../adapt/gro-adapter-api-server.js')).create_adapter() : null,
+			enable_api_server
+				? (await import('../adapt/gro-adapter-generic-build.js')).create_adapter({
+						build_name: API_SERVER_BUILD_CONFIG.name,
+				  })
+				: null,
 			// enable_gro_frontend
 			// 	? (await import('../adapt/gro-adapter-spa-frontend.js')).create_adapter()
 			// 	: null,